moved all objects and implementations to their own .h/.cpp files, related to #24
