+++ hash_table.h
@@ -1,4 +1,4 @@
-#include"user.h"
+#include"list.h"
 
 //Samuel Parker, 10/2019
